fixed line drawing

+++ LinesDetection/src/includes/image_operations.h
@@ -16,8 +16,8 @@
 /*
  * Adds lines to image, to visualize.
  */
-extern void add_lines_to_image(image_t image, lines_t lines, uint8_t color_line,
-        uint8_t color_bg);
+extern void add_lines_to_image(image_t image, lines_t lines, int color_line,
+        int color_bg);
 
 /*
  * New Gaussian kernel.
